requirements/local.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-smart-selects==1.5.9
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
Unipath==1.1